SRR-111-stock-service-and-repository-implemented-and-repository-package-moved
